Refactoring duplicate firebase functions
extracted the functions from the components and stored them into FirebaseFunctions that can be used anywhere
