+++ apps/native/components/FirebaseFunctions.tsx
@@ -1,0 +1,29 @@
+import {doc, getDoc} from "firebase/firestore";
+import {db} from "@native/firebase";
+
+export const fetchUserRole = async (uid: string) => {
+    const docRef = doc(db, "users", uid);
+    const docSnap = await getDoc(docRef);
+
+    if (docSnap.exists()) {
+        console.log("Document data:", docSnap.data());
+        return docSnap.data().role;
+    } else {
+        // docSnap.data() will be undefined in this case
+        return "notFound";
+    }
+};
+
+export const fetchUsername = async (uid: string) => {
+    const docRef = doc(db, "users", uid);
+    const docSnap = await getDoc(docRef);
+
+    if (docSnap.exists()) {
+        console.log("Document data:", docSnap.data());
+        return docSnap.data().username;
+    } else {
+        // docSnap.data() will be undefined in this case
+        console.log("No such document!");
+        return "notFound";
+    }
+}
