+++ apps/native/screens/SignupScreen.tsx
@@ -20,6 +20,8 @@ import FontAwesome5 from 'react-native-vector-icons/FontAwesome5';
 import { createUserWithEmailAndPassword } from '@firebase/auth';
 import { setDoc, doc, getDoc } from 'firebase/firestore';
 import { signInWithEmailAndPassword } from 'firebase/auth';
+import {fetchUserRole} from "@native/components/FirebaseFunctions";
+
 
 const SignupScreen = ({ navigation }: any) => {
     const [email, setEmail] = useState('');
@@ -31,18 +33,6 @@ const SignupScreen = ({ navigation }: any) => {
     const [open, setOpen] = useState(false);
 
     const handleSignup = async () => {
-        async function returnRole(uid: string) {
-            const docRef = doc(db, "users", uid);
-            const docSnap = await getDoc(docRef);
-
-            if (docSnap.exists()) {
-                console.log("Document data:", docSnap.data());
-                return docSnap.data().role;
-            } else {
-                // docSnap.data() will be undefined in this case
-                return "notFound";
-            }
-        }
 
         try {
             const userCredential = await createUserWithEmailAndPassword(auth, email, password);
@@ -55,7 +45,7 @@ const SignupScreen = ({ navigation }: any) => {
 
             const signInCredential = await signInWithEmailAndPassword(auth, email, password);
             await AsyncStorage.setItem('userUID', signInCredential.user.uid);
-            await AsyncStorage.setItem('userRole', await returnRole(signInCredential.user.uid));
+            await AsyncStorage.setItem('userRole', await fetchUserRole(signInCredential.user.uid));
             setConnectionStatus("Success!");
             navigation.navigate('UserProfile');
         } catch (error) {
@@ -72,11 +62,11 @@ const SignupScreen = ({ navigation }: any) => {
             <TouchableWithoutFeedback>
                 <View style={styles.flexibleContainer}>
                     <View style={styles.header}>
-                        <TouchableOpacity onPress={() => navigation.goBack()} style={styles.headerIcon}>
+                        <TouchableOpacity id={"goBackBtn"} onPress={() => navigation.goBack()} style={styles.headerIcon}>
                             <FontAwesome5 name="times" solid size={24} color="#000" />
                         </TouchableOpacity>
                         <Text style={styles.headerTitle}>Signup</Text>
-                        <TouchableOpacity onPress={() => navigation.navigate('Login')} style={styles.headerRightText}>
+                        <TouchableOpacity id={"loginInBtn"} onPress={() => navigation.navigate('Login')} style={styles.headerRightText}>
                             <Text style={styles.headerText}>Log In</Text>
                         </TouchableOpacity>
                     </View>
@@ -101,6 +91,7 @@ const SignupScreen = ({ navigation }: any) => {
                             />
 
                             <DropDownPicker
+                                testID={"roleDropDownPicker"}
                                 open={open}
                                 setOpen={setOpen}
                                 value={role}
@@ -121,6 +112,7 @@ const SignupScreen = ({ navigation }: any) => {
                                     autoCapitalize="none"
                                 />
                                 <TouchableOpacity
+                                    id={"showSignupPasswordBtn"}
                                     onPress={() => setPasswordVisible(!passwordVisible)}
                                     style={styles.showButton}
                                 >
